+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - snap package not starting on Wayland
+- break windows not closing correctly on all platforms
 
 ## [1.19.0] - 2025-11-13
 ### Added